feat: Improved docs and made swagger doc optional
- Fixed the hml file for the plugin
- updated README.md
- code fixes to make swagger doc optional
